WeightedInvertedFile

InvertedFiles.WeightedInvertedFile — Type

struct WeightedInvertedFile <: AbstractInvertedFile

An inverted index is a sparse matrix representation of with floating point weights, it supports only positive non-zero values. This index is optimized to efficiently solve k nearest neighbors (cosine distance, using previously normalized vectors).

Parameters

lists: posting lists (non-zero id-elements in rows)

weights: non-zero weights (in rows)

sizes: number of non-zero values in each element (non-zero values in columns)

locks: per-row locks for multithreaded construction

BinaryInvertedFile

InvertedFiles.BinaryInvertedFile — Type

struct BinaryInvertedFile <: AbstractInvertedFile

Creates a binary weighted inverted index. An inverted index is an sparse matrix representation optimized for computing k nn elements (columns) under some distance.

Properties:

dist: Distance function to be applied, valid values are: IntersectionDissimilarity(), DiceDistance(), JaccardDistance(), and `CosineDistanceSet()

lists: posting lists (non-zero values of the rows in the matrix)

sizes: number of non-zero values per object (number of non-zero values per column)

locks: Per row locks for multithreaded construction

Searching algorithms

InvertedFiles.prepare_posting_lists_for_querying — Function

prepare_posting_lists_for_querying(idx::AbstractInvertedFile, q, pools=getpools(idx), tol=1e-6)

Fetches and prepares the involved posting lists to solve q

SimilaritySearch.search — Function

search(idx::AbstractInvertedFile, q, res::KnnResult; pools=nothing)

Searches q in idx using the cosine dissimilarity, it computes the full operation on idx. res specify the query

search(callback::Function, idx::WeightedInvertedFile, Q, P; t=1)

Find candidates for solving query Q using idx. It calls callback on each candidate (objID, dist)

Arguments:

callback: callback function on each candidate

idx: inverted index

Q: the set of involved posting lists, see prepare_posting_lists_for_querying

P: a vector of starting positions in Q (initial state as ones)

search(callback::Function, idx::BinaryInvertedFile, Q, P, t)

Find candidates for solving query Q using idx. It calls callback on each candidate (objID, dist)

Arguments

callback: callback function on each candidate

idx: inverted index

Q: the set of involved posting lists, see prepare_posting_lists_for_querying

P: a vector of starting positions in Q (initial state as ones)

t: threshold (t=1 union, t > 1 solves the t-threshold problem)

Sparse matrices

Inverted indexes/files are representations of sparse matrices optimized for certain operations. We provide some functions to convert inverted files to sparse matrices.

SparseArrays.sparse — Function

sparse(idx::BinaryInvertedFile, one::Type{RealType}=1f0)

Creates an sparse matrix (from SparseArrays) from idx using one as value.

   I  
   ↓    1 2 3 4 5 … n  ← J
 L[1] = 0 1 0 0 1 … 0
 L[2] = 1 0 0 1 0 … 1
 L[3] = 1 0 1 0 0 … 1
 ⋮
 L[m] = 0 0 1 1 0 … 0

sparse(idx::WeightedInvertedFile)

Creates an sparse matrix (from SparseArrays) from idx

SparseArrays.sparsevec — Function

sparsevec(vec::DVEC{Ti,Tv}, m=0) where {Ti<:Integer,Tv<:Number}

Creates a sparse vector from a DVEC sparse vector

Inverted indexes constructors also support sparse matrices as input (wrapped on MatrixDatabase structs)

Dictionary-based sparse vectors

Some application domains could take advantage of hash based sparse vectors, and some of them are the target of InvertedFiles, therefore, the package also provide a partial implementation of sparse vectors using Dict.

InvertedFiles.dvec — Function

dvec(x::AbstractSparseVector)

Converts an sparse vector into a DVEC sparse vector

Missing docstring.

Missing docstring for DVEC. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SVEC. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SVEC32. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SVEC64. Check Documenter's build log for details.

Missing docstring.

Missing docstring for nnz. Check Documenter's build log for details.

Missing docstring.

Missing docstring for findmax. Check Documenter's build log for details.

Missing docstring.

Missing docstring for argmax. Check Documenter's build log for details.

Missing docstring.

Missing docstring for maximum. Check Documenter's build log for details.

Missing docstring.

Missing docstring for findmin. Check Documenter's build log for details.

Missing docstring.

Missing docstring for argmin. Check Documenter's build log for details.

Missing docstring.

Missing docstring for minimum. Check Documenter's build log for details.

LinearAlgebra.normalize! — Function

normalize!(bow::DVEC)

Inplace normalization of bow

LinearAlgebra.dot — Function

dot(a::DVEC, b::DVEC)::Float64

Computes the dot product for two DVEC vectors

LinearAlgebra.norm — Function

norm(a::DVEC)::Float64

Computes a normalized DVEC vector

Base.zero — Function

zero(::Type{DVEC{Ti,Tv}}) where {Ti,Tv}

Creates an empty DVEC vector

InvertedFiles.add! — Function

add!(a::DVEC{Ti,Tv}, b::DVEC{Ti,Tv}) where {Ti,Tv<:Real}
add!(a::DVEC{Ti,Tv}, b::AbstractSparseArray) where {Ti,Tv<:Real}
add!(a::DVEC{Ti,Tv}, b::Pair{Ti,Tv}) where {Ti,Tv<:Real}

Updates a to the sum of a+b

Base.sum — Function

Base.sum(col::AbstractVector{<:DVEC})

Computes the sum of the given list of vectors

Base.:+ — Function

+(a::DVEC{Ti,Tv}, b::DVEC{Ti,Tv}) where {Ti,Tv<:Real}
+(a::DVEC, b::Pair)

Computes the sum of a and b

Base.:- — Function

-(a::DVEC{Ti,Tv}, b::DVEC{Ti,Tv}) where {Ti,Tv<:Real}

Substracts of b of a

Base.:* — Function

*(a::DVEC{Ti,Tv}, b::DVEC{Ti,Tv}) where {Ti,Tv<:Real}
*(a::DVEC{K, V}, b::F) where K where {V<:Real} where {F<:Real}

Computes the element-wise product of a and b

Base.:/ — Function

/(a::DVEC{K, V}, b::F) where K where {V<:Real} where {F<:Real}

Computes the element-wise division of a and b

InvertedFiles.centroid — Function

centroid(cluster::AbstractVector{<:DVEC})

Computes a centroid of the given list of DVEC vectors

Distances.evaluate — Function

evaluate(::NormalizedCosineDistance, a::DVEC, b::DVEC)::Float64

Computes the cosine distance between two DVEC sparse vectors

It supposes that bags are normalized (see normalize! function)

evaluate(::CosineDistance, a::DVEC, b::DVEC)::Float64

Computes the cosine distance between two DVEC sparse vectors

evaluate(::NormalizedAngleDistance, a::DVEC, b::DVEC)::Float64

Computes the angle between two DVEC sparse vectors

It supposes that all bags are normalized (see normalize! function)

evaluate(::AngleDistance, a::DVEC, b::DVEC)::Float64

Computes the angle between two DVEC sparse vectors

Missing docstring.

Missing docstring for NormalizedAngleDistance. Check Documenter's build log for details.

Missing docstring.

Missing docstring for NormalizedCosineDistance. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AngleDistance. Check Documenter's build log for details.

Missing docstring.

Missing docstring for CosineDistance. Check Documenter's build log for details.

Missing docstring.

Missing docstring for evaluate. Check Documenter's build log for details.