Inverted files

Missing docstring.

Missing docstring for append!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for push!. Check Documenter's build log for details.

InvertedFiles.prepare_posting_lists_for_querying — Function

prepare_posting_lists_for_querying(idx::AbstractInvertedFile, q, tol, pools)

Fetches and prepares the involved posting lists to solve q

Missing docstring.

Missing docstring for search!. Check Documenter's build log for details.

WeightedInvertedFile

InvertedFiles.WeightedInvertedFile — Type

struct WeightedInvertedFile <: AbstractInvertedFile

An inverted index is a sparse matrix representation of with floating point weights, it supports only positive non-zero values. This index is optimized to efficiently solve k nearest neighbors (cosine distance, using previously normalized vectors).

Parameters

lists: posting lists (non-zero id-elements in rows)

weights: non-zero weights (in rows)

sizes: number of non-zero values in each element (non-zero values in columns)

BinaryInvertedFile

InvertedFiles.BinaryInvertedFile — Type

struct BinaryInvertedFile <: AbstractInvertedFile

Creates a binary weighted inverted index. An inverted index is an sparse matrix representation optimized for computing k nn elements (columns) under some distance.

Properties:

dist: Distance function to be applied, valid values are: IntersectionDissimilarity(), DiceDistance(), JaccardDistance(), and `CosineDistanceSet()

lists: posting lists (non-zero values of the rows in the matrix)

sizes: number of non-zero values per object (number of non-zero values per column)

locks: Per row locks for multithreaded construction

Sparse matrices

Inverted indexes/files are representations of sparse matrices optimized for certain operations. We provide some functions to convert inverted files to sparse matrices.

Missing docstring.

Missing docstring for sparse. Check Documenter's build log for details.

Missing docstring.

Missing docstring for sparsevec. Check Documenter's build log for details.

Inverted indexes constructors also support sparse matrices as input (wrapped on MatrixDatabase structs)